WEAT-9 impl network logic for weekly weather

+++ core/domain/src/main/java/io/wookoo/domain/model/weather/current/CurrentDayModel.kt
@@ -2,13 +2,13 @@ package io.wookoo.domain.model.weather.current
 
 data class CurrentDayModel(
     val time: Long,
-    val temperature: Float,
+    val temperature: Double,
     val relativeHumidity: Int,
-    val feelsLike: Float,
+    val feelsLike: Double,
     val isDay: Boolean,
     val precipitation: PrecipitationModel,
     val cloudCover: Int,
-    val pressureMSL: Float,
+    val pressureMSL: Double,
     val wind: WindModel,
     val weatherStatus: Int,
 )
