+++ core/domain/src/main/java/io/wookoo/domain/model/weather/current/WindModel.kt
@@ -2,6 +2,6 @@ package io.wookoo.domain.model.weather.current
 
 data class WindModel(
     val direction: Int,
-    val speed: Float,
-    val gust: Float,
+    val speed: Double,
+    val gust: Double,
 )
