+++ core/network/src/main/java/io/wookoo/network/dto/weather/current/CurrentWeatherDto.kt
@@ -9,31 +9,31 @@ data class CurrentWeatherDto(
     @SerialName("time")
     val time: Long,
     @SerialName("temperature_2m")
-    val temperature: Float,
+    val temperature: Double,
     @SerialName("relative_humidity_2m")
     val relativeHumidity: Int,
     @SerialName("apparent_temperature")
-    val feelsLike: Float,
+    val feelsLike: Double,
     @SerialName("is_day")
     val isDay: Int,
     @SerialName("precipitation")
-    val precipitation: Float,
+    val precipitation: Double,
     @SerialName("rain")
-    val rain: Float,
+    val rain: Double,
     @SerialName("showers")
-    val showers: Float,
+    val showers: Double,
     @SerialName("snowfall")
-    val snowfall: Float,
+    val snowfall: Double,
     @SerialName("cloud_cover")
     val cloudCover: Int,
     @SerialName("pressure_msl")
-    val pressureMSL: Float,
+    val pressureMSL: Double,
     @SerialName("wind_direction_10m")
     val windDirection: Int,
     @SerialName("wind_speed_10m")
-    val windSpeed: Float,
+    val windSpeed: Double,
     @SerialName("wind_gusts_10m")
-    val windGusts: Float,
+    val windGusts: Double,
     @SerialName("weather_code")
     val weatherCode: Int,
 )
